Viha/bug fixes (#94)
* fix single-order bug; clean up code

* fix ts

* decimal fix

* routes

* refactored useEffects and fixed image issues

* Signup Dup Emails

Error for dup emails / IMAGE config

* vehicle management bug fixes

* copilot fixes

* Sign Up Value Persistance

If error, then Values stay in fields

* Login Error Message

Fixing the Error messages for Login. More info

* Maintain Login Values

* Live Password Requirement Checker

Sign Up form checks if zod requirements are met.

* 20lb >= Fee

If weight is 20lbs or more than fee.

* Logged In Cant visit Login

If the User is logged in then they cannot visit the log in page.

* Protect Admin

This changes added a server action to protect admin pages. Each page and the layout is protected.

* added proper image url validation, data retainment on form errors

* Admin / EMPL fix through routing

* transaction race condition fix

---------

Co-authored-by: leo-kildani <kildani.leonardo@gmail.com>
Co-authored-by: vihashah <viha.shah@sjsu.edu>
Co-authored-by: Ausxh <ausxh66@gmail.com>